refactor: load settings panes on demand and drop the metadata split
The command palette listed the built-in panes array, so a pane registered by
an extension showed in the settings dialog but not in the palette.

Load each pane with defineAsyncComponent. The registry then holds no pane code,
so the palette can read the registry itself. This also removes the reason for
the separate metadata module, so panes.ts folds back into index.ts.

Ranks go away with it: the registry falls back to registration order, and the
array is already in display order.

The settings chunk drops from 1.1 MB to 2.9 KB. Each pane now loads when the
user opens it.
